requirements.txt
datetime
plyer
ttkbootstrap
Pillow
pygame
pystray